vendor: copy go-eth crypto files to vendor
Because dep doesn't grab the C header files necessary
for building, so they have to be copied manually.

This sucks, but they gotta be there.

Instead of using rsync to recursively copy these
files from their original source paths under GOPATH,
you could go in the vendor copy and git clone the
repo, but that's IMO exactly as ugly.